requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.32.5
Pillow==10.0.1
django-admin-lte3==0.1.6
django-admin-lte3-theme==0.1.3
pytz==2025.2